Support "data__" props on Gtk components
Ok, so this one makes me feel yuck inside.  Basically you have a
Gtk.ListBoxRow, which is activatable.  This means you get a signal
"row-activate" (or something like that) on the parent list.  So you need
to be able at attach a python attribute to the child element.  Therefore
this ugly prop.  It would be nicer if you could just get a signal on the
children, but whatever.